[CSL-860] Sort out some stuff in Pos.Binary
