+++ src/Pos/Binary/Communication.hs
@@ -26,7 +26,6 @@ import           Pos.Delegation.Types             (ConfirmProxySK (..), SendProx
 import           Pos.DHT.Model.Types              (meaningPartLength)
 import           Pos.Ssc.Class.Helpers            (SscHelpersClass)
 import           Pos.Ssc.Class.Types              (Ssc (..))
-import           Pos.Txp.Network.Types            (TxMsgTag (..))
 import           Pos.Update.Network.Types         (ProposalMsgTag (..), VoteMsgTag (..))
 
 
@@ -37,68 +36,8 @@ import           Pos.Update.Network.Types         (ProposalMsgTag (..), VoteMsgT
 deriving instance Bi MessageName
 
 ----------------------------------------------------------------------------
--- Blocks
+-- Protocol version info and related
 ----------------------------------------------------------------------------
-
-instance Bi MsgGetHeaders where
-    put (MsgGetHeaders f t) = put f >> put t
-    get = label "MsgGetHeaders" $ MsgGetHeaders <$> get <*> get
-
-instance Bi MsgGetBlocks where
-    put (MsgGetBlocks f t) = put f >> put t
-    get = label "MsgGetBlocks" $ MsgGetBlocks <$> get <*> get
-
-instance Ssc ssc => Bi (MsgHeaders ssc) where
-    put (MsgHeaders b) = put b
-    get = label "MsgHeaders" $ MsgHeaders <$> get
-
-instance SscHelpersClass ssc => Bi (MsgBlock ssc) where
-    put (MsgBlock b) = put b
-    get = label "MsgBlock" $ MsgBlock <$> get
-
-----------------------------------------------------------------------------
--- Transaction processing
-----------------------------------------------------------------------------
-
-instance Bi TxMsgTag where
-    put TxMsgTag = pure ()
-    get = pure TxMsgTag
-
-----------------------------------------------------------------------------
--- Delegation/PSK
-----------------------------------------------------------------------------
-
-instance Bi SendProxySK where
-    put (SendProxySKLight pSk) = putWord8 0 >> put pSk
-    put (SendProxySKHeavy pSk) = putWord8 1 >> put pSk
-    get = label "SendProxySK" $ getWord8 >>= \case
-        0 -> SendProxySKLight <$> get
-        1 -> SendProxySKHeavy <$> get
-        t -> fail $ "get@SendProxySK: unknown tag " <> show t
-
-instance Bi ConfirmProxySK where
-    put (ConfirmProxySK pSk proof) = put pSk >> put proof
-    get = label "ConfirmProxySK" $ liftA2 ConfirmProxySK get get
-
---instance Bi CheckProxySKConfirmed where
---    put (CheckProxySKConfirmed pSk) = put pSk
---    get = CheckProxySKConfirmed <$> get
---
---instance Bi CheckProxySKConfirmedRes where
---    put (CheckProxySKConfirmedRes res) = put res
---    get = CheckProxySKConfirmedRes <$> get
---
-----------------------------------------------------------------------------
--- Update system
-----------------------------------------------------------------------------
-
-instance Bi ProposalMsgTag where
-    put ProposalMsgTag = pure ()
-    get = pure ProposalMsgTag
-
-instance Bi VoteMsgTag where
-    put VoteMsgTag = pure ()
-    get = pure VoteMsgTag
 
 -- Encoding of HandlerSpec is as follow:
 --
@@ -141,3 +80,61 @@ instance Bi PeerId where
                         then error $ sformat ("Wrong PeerId length "%int) (BS.length b)
                         else putByteString b
     get = label "PeerId" $ PeerId <$> getByteString peerIdLength
+
+-- TODO: move into each component
+
+----------------------------------------------------------------------------
+-- Blocks
+----------------------------------------------------------------------------
+
+instance Bi MsgGetHeaders where
+    put (MsgGetHeaders f t) = put f >> put t
+    get = label "MsgGetHeaders" $ MsgGetHeaders <$> get <*> get
+
+instance Bi MsgGetBlocks where
+    put (MsgGetBlocks f t) = put f >> put t
+    get = label "MsgGetBlocks" $ MsgGetBlocks <$> get <*> get
+
+instance Ssc ssc => Bi (MsgHeaders ssc) where
+    put (MsgHeaders b) = put b
+    get = label "MsgHeaders" $ MsgHeaders <$> get
+
+instance SscHelpersClass ssc => Bi (MsgBlock ssc) where
+    put (MsgBlock b) = put b
+    get = label "MsgBlock" $ MsgBlock <$> get
+
+----------------------------------------------------------------------------
+-- Delegation/PSK
+----------------------------------------------------------------------------
+
+instance Bi SendProxySK where
+    put (SendProxySKLight pSk) = putWord8 0 >> put pSk
+    put (SendProxySKHeavy pSk) = putWord8 1 >> put pSk
+    get = label "SendProxySK" $ getWord8 >>= \case
+        0 -> SendProxySKLight <$> get
+        1 -> SendProxySKHeavy <$> get
+        t -> fail $ "get@SendProxySK: unknown tag " <> show t
+
+instance Bi ConfirmProxySK where
+    put (ConfirmProxySK pSk proof) = put pSk >> put proof
+    get = label "ConfirmProxySK" $ liftA2 ConfirmProxySK get get
+
+--instance Bi CheckProxySKConfirmed where
+--    put (CheckProxySKConfirmed pSk) = put pSk
+--    get = CheckProxySKConfirmed <$> get
+--
+--instance Bi CheckProxySKConfirmedRes where
+--    put (CheckProxySKConfirmedRes res) = put res
+--    get = CheckProxySKConfirmedRes <$> get
+
+----------------------------------------------------------------------------
+-- Update system
+----------------------------------------------------------------------------
+
+instance Bi ProposalMsgTag where
+    put ProposalMsgTag = pure ()
+    get = pure ProposalMsgTag
+
+instance Bi VoteMsgTag where
+    put VoteMsgTag = pure ()
+    get = pure VoteMsgTag
