+++ src/Pos/Txp/Core.hs
@@ -1,3 +1,5 @@
 -- | Basic functionality of Txp.
 
 {-# OPTIONS_GHC -F -pgmF autoexporter #-}
+
+import           Pos.Binary.Txp.Core ()
