+++ src/Pos/Txp/Network/Types.hs
@@ -1,4 +1,3 @@
-
 -- | Types used for communication about Blocks.
 
 module Pos.Txp.Network.Types
@@ -6,13 +5,12 @@ module Pos.Txp.Network.Types
        , TxMsgContents (..)
        ) where
 
-import qualified Data.Text.Buildable     as Buildable
-import           Formatting              (bprint, build, (%))
+import qualified Data.Text.Buildable as Buildable
+import           Formatting          (bprint, build, (%))
 import           Universum
 
-import           Pos.Binary.Core.Address ()
-import           Pos.Binary.Txp          ()
-import           Pos.Txp.Core.Types      (Tx, TxDistribution, TxWitness)
+import           Pos.Binary.Core     ()
+import           Pos.Txp.Core        (Tx, TxDistribution, TxWitness)
 
 data TxMsgTag = TxMsgTag deriving (Eq, Show)
